requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.7.4
Django==5.0.7
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-redis==5.4.0
djangorestframework==3.15.2
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
packaging==24.1
python-dateutil==2.9.0.post0
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
